+++ ui/packages/app/src/components/input/Image3D/config.ts
@@ -1,4 +1,4 @@
-import Component from "./Model3d.svelte";
+import Component from "./Image3D.svelte";
 import ExampleComponent from "./Example.svelte";
 import { loadAsData } from "../../utils/example_processors";
